Introduce targets without setuid
The reason for this that we would like to give Ports users an option to
install intel_backlight without the setuid bit set. I was suggested to
look at the x11-servers/xorg-server port as an example to follow.

The Xorg server port defines an option called SUID, which controls
whether the final bin/Xorg binary has a setuid bit set or not. The port
is built without any setuid bits set, which is good because it does not
cause any security warnings. It is the pkg-plist, where setuid setting
is handled.

In order to follow the example of the Xorg server port we have to be
able to build port without any setuids set (hence I replaced the 4555
permissions with 0555 in the install target). The setuid will be set in
pkg-plist. I think that those changes are reasonable as this repository
is maintained primarily for FreeBSD Ports. Nevertheless, I introduced
special setuid targets, namely, install-setuid and install-setuid-strip,
which are for those users, who install intel_backlight manually from
the repository.